docs: Break Styleguide stories into separate files
